Cleaned up code and added more comments
